requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# Groq for Llama 3.3 70B
langchain-groq>=0.0.1

# Vector Store & Embeddings (for Memory Agent)
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# LangGraph (for multi-agent orchestration)
langgraph>=0.0.20

# Utilities
python-dotenv>=1.0.0

# Optional but recommended
pydantic>=2.0.0


streamlit>=1.28.0
python-dotenv>=1.0.0
groq>=0.4.0
langchain>=0.1.0
langchain-groq>=0.0.1
chromadb>=0.4.0
sentence-transformers>=2.2.0

streamlit>=1.28.0
python-dotenv>=1.0.0
groq>=0.4.0
langchain>=0.1.0
langchain-groq>=0.0.1
chromadb>=0.4.0
sentence-transformers>=2.2.0